% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_se.R
\name{add_contaminants}
\alias{add_contaminants}
\title{Adds a Potential.contaminant column to pg_matrix based on MaxQuant
contaminants.txt}
\usage{
add_contaminants(pg_matrix)
}
\arguments{
\item{pg_matrix}{The report.pg_matrix output file from DIANN}
}
\value{
A data frame with a added Potential.contaminant column.
}
\description{
Adds a Potential.contaminant column to pg_matrix based on MaxQuant
contaminants.txt
}
\examples{
pg_matrix <- add_contaminants(report.pg_matrix)
}
