% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{add_iBAQ}
\alias{add_iBAQ}
\title{Add iBAQ intensities to report.pg_matrix file.}
\usage{
add_iBAQ(pg_matrix, pr_matrix, ibaq_stats = NULL, organism = "hs")
}
\arguments{
\item{pg_matrix}{The report.pg_matrix file.}

\item{pr_matrix}{The report.pr_matrix file.}

\item{ibaq_stats}{Dataframe with proteinIDs and associated number of iBAQ peptides obtained with \link{get_ibaq_peptides}.
Required when not using pre-calculated iBAQ peptides}

\item{organism}{Specifies which organism to use if using pre-calculated iBAQ peptides. 'hs' for human, 'mm' for mouse.}
}
\value{
A report.pg_matrix data frame with iBAQ columns added.
}
\description{
Add iBAQ intensities to report.pg_matrix file.
}
\examples{
pg <- add_iBAQ(report.pg_matrix, report.pr_matrix, organism = 'hs')
}
