% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_se.R
\name{add_maxLFQ_iBAQ}
\alias{add_maxLFQ_iBAQ}
\title{Adds iBAQ values based on LFQ values to SummarizedExperiment}
\usage{
add_maxLFQ_iBAQ(se)
}
\arguments{
\item{se}{SummarizedExperiment object in prepare_se}
}
\value{
A Summarized Experiment object with extra assay.
}
\description{
Adds iBAQ values based on LFQ values to SummarizedExperiment
}
\examples{
\dontrun{
se <- add_maxLFQ_iBAQ(se)
}
}
