% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_functions.R
\name{add_median_peptide_intensity}
\alias{add_median_peptide_intensity}
\title{Adds median peptide intensities to summerizedExperiment object}
\usage{
add_median_peptide_intensity(se)
}
\arguments{
\item{se}{SummerizedExperiment object returned from DEP::make_se()}
}
\value{
a summerizedExperiment object with peptide intensities as extra assay
}
\description{
Adds median peptide intensities to summerizedExperiment object
}
\examples{

se <- prepare_se(report.pg_matrix, expDesign, report.pr_matrix)
se <- add_median_peptide_intensity(se)
}
