% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_functions.R
\name{add_peptide_numbers}
\alias{add_peptide_numbers}
\title{Add the peptide number information to pg_matrix}
\usage{
add_peptide_numbers(pg_matrix, peptide_numbers, id_column = "Protein.Group")
}
\arguments{
\item{pg_matrix}{The report.pg_matrix file}

\item{peptide_numbers}{The output from \link{get_nPep_prMatrix}}

\item{id_column}{The column in pr_matrix that is used to match the
peptide_numbers and pg_matrix. Should be identical to the column used in
get_nPEP_prMatrix}
}
\value{
a pg_matrix data frame with added peptide number information
}
\description{
Add the peptide number information to pg_matrix
}
\examples{
peptide_numbers <- get_nPep_prMatrix(report.pr_matrix)
pg_matrix <- add_peptide_numbers(report.pg_matrix, peptide_numbers)
}
