% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_standardTheme}
\alias{add_standardTheme}
\title{Adds standard changes to ggplot theme}
\usage{
add_standardTheme(plot)
}
\arguments{
\item{plot}{a ggplot object}
}
\value{
a ggplot object
}
\description{
Adds standard changes to ggplot theme
}
\examples{
library(ggplot2)
df <- data.frame(var1 = 1:10, var2 = sample(1:100, 10))
p <- ggplot(df, aes(x = var1, y = var2)) + geom_point()
p <- add_standardTheme(p)
}
