% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{calculate_iBAQ}
\alias{calculate_iBAQ}
\title{Calculate iBAQ values from raw intensities}
\usage{
calculate_iBAQ(pr_matrix, ibaq_stats = NULL, organism = "hs")
}
\arguments{
\item{pr_matrix}{The report.pr_matrix file}

\item{ibaq_stats}{Dataframe with proteinIDs and associated number of iBAQ peptides obtained with \link{get_ibaq_peptides}.
Required when not using pre-calculated iBAQ peptides}

\item{organism}{Specifies which organism to use if using pre-calculated iBAQ peptides. 'hs' for human, 'mm' for mouse.}
}
\value{
A dataframe with iBAQ values per proteinGroup per sample and the number
of iBAQ peptides.
}
\description{
Calculate iBAQ values from raw intensities
}
\examples{
ibaq_values <- calculate_iBAQ(report.pr_matrix, organism = 'hs') # For human
ibaq_values <- calculate_iBAQ(report.pr_matrix, organism = 'mm') # For mouse
}
