% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contaminants_maxquant}
\alias{contaminants_maxquant}
\title{Data frame with potential contaminants from MaxQuant.}
\format{
\subsection{\code{contaminants_maxquant}}{

A data frame with 246 rows and 1 columns.
}
}
\source{
MaxQuant
}
\usage{
contaminants_maxquant
}
\description{
Data frame with potential contaminants from MaxQuant.
}
\keyword{datasets}
