% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expDesign}
\alias{expDesign}
\title{Experimental design for example data}
\format{
\subsection{\code{expDesign}}{

A data frame with 6 rows and 3 columns:
\describe{
\item{label}{column name in report.pg_matrix specifying MS sample}
\item{condition}{The different experimental conditions}
\item{replicate}{The different replicates per condition}
}
}
}
\source{
241220_V118_LK_DNA_PD
}
\usage{
expDesign
}
\description{
Necessary for creating SummarizedExperiment object with MS data
}
\keyword{datasets}
