% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP.R
\name{get_DEPresults}
\alias{get_DEPresults}
\title{Perform differential protein expression analysis}
\usage{
get_DEPresults(
  se,
  condition1 = NULL,
  condition2 = NULL,
  ref_condition = NULL,
  tests = NULL,
  alpha = 0.05,
  lfc = 1,
  type = "manual",
  fdr_type = "BH"
)
}
\arguments{
\item{se}{SummarizedExperiment object with protein expression data.}

\item{condition1}{One of the conditions to be tested present in expDesign.}

\item{condition2}{One of the conditions to be tested present in expDesign.}

\item{ref_condition}{One of the conditions in expDesign that is used as
reference. Must be used in combination with type = 'control'.}

\item{tests}{Character vector specifying which contrasts to test when using
"type = 'manual'"}

\item{alpha}{the p.adjust significance cutoff.}

\item{lfc}{The log2 fold change significance cutoff.}

\item{type}{Type of comparison to be made. Options are: 'manual', 'control',
and 'all'.}

\item{fdr_type}{Type of fdr correction. Options are 'fdrtool' and
'BH' (Benjamini-Hochberg). Default is 'BH'}
}
\value{
A data frame with the results of differential protein expression
analysis
}
\description{
Perform differential protein expression analysis
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, 'motif1', 'neg_ctrl', type = 'manual')
}
