% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_imputation.R
\name{get_detection_limit}
\alias{get_detection_limit}
\title{Find the cutoff for MAR/MNAR classification}
\usage{
get_detection_limit(data)
}
\arguments{
\item{data}{A data matrix with replicates of a single condition}
}
\value{
A numeric value
}
\description{
Find the cutoff for MAR/MNAR classification
}
\examples{
\dontrun{
cutoff <- get_detection_limit(data)
}
}
