% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{get_genesets}
\alias{get_genesets}
\title{Extract a collection of gene sets from the MSigDB object}
\usage{
get_genesets(
  db,
  collection = c("c2", "c5", "h"),
  subcollection = c("GOBP", "GOCC", "GOMF", "KEGG")
)
}
\arguments{
\item{db}{An MSigDB object downloaded with load_msigdb()}

\item{collection}{Specifies which collection of gene sets to use. Options are: c2 (e.g. Reactome, KEGG), c5 (GO sets), or h (cancer hallmarks)}

\item{subcollection}{Specifies which subcollection to use. E.g. 'Reactome' or 'KEGG' for c2, or 'GOBP/GOCC/GOMF' for 'c5'}
}
\value{
A list with gene sets
}
\description{
Extract a collection of gene sets from the MSigDB object
}
\examples{
\dontrun{
Reactome <- get_genesets(db, collection = 'c2', subcollection = 'Reactome')
}

}
