% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{get_ibaq_peptides}
\alias{get_ibaq_peptides}
\title{Perform in silico tryptic digest on uniprot fasta file}
\usage{
get_ibaq_peptides(fasta_location, miscleaves = 0, enzyme = "trypsin")
}
\arguments{
\item{fasta_location}{Path to fasta file}

\item{miscleaves}{Allowed number of miscleavages. Defaults to 0 (standard for iBAQ)}

\item{enzyme}{The enzyme used for digestion. Defaults to 'trypsin'. See 'cleaver'
package vignette for more options.}
}
\value{
A data frame with the number of theoretically observable peptides for
each entry in the fasta file.
}
\description{
Perform in silico tryptic digest on uniprot fasta file
}
\examples{
\dontrun{
no_ibaq_peptides <- get_ibaq_peptides('path/to/fasta.fasta')
}
}
