% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_functions.R
\name{get_intensities_prMatrix}
\alias{get_intensities_prMatrix}
\title{Calculate protein intensities from peptide information}
\usage{
get_intensities_prMatrix(
  pr_matrix,
  id_column = "Protein.Group",
  peptide_level = "Stripped.Sequence"
)
}
\arguments{
\item{pr_matrix}{The report.pr_matrix file}

\item{id_column}{The IDs used in the output file}

\item{peptide_level}{he column specifying at which level intensities are
aggregated. See \link{summarize_peptide_intensities}}
}
\value{
A matrix with the summed peptides intenties of razor/unique peptides
}
\description{
Calculate protein intensities from peptide information
}
\examples{
mat <- get_intensities_prMatrix(report.pr_matrix)
}
