% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_functions.R
\name{get_median_intensities_prMatrix}
\alias{get_median_intensities_prMatrix}
\title{Calculates the median peptide intensity per proteinGroup per sample.}
\usage{
get_median_intensities_prMatrix(
  pr_matrix,
  id_column = "Protein.Group",
  peptide_level = "Stripped.Sequence"
)
}
\arguments{
\item{pr_matrix}{The report.pr_matrix file}

\item{id_column}{The column in pr_matrix that is used as identifier in the
output file}

\item{peptide_level}{The column in pr_matrix that is used to aggregate
peptide data}
}
\value{
a matrix with median peptide intensities per sample per proteinGroup
}
\description{
The median peptide intensity (MPI) can be used instead of iBAQ as proxy
for protein abundancy.
}
\examples{
mpi <- get_median_intensities_prMatrix(report.pr_matrix)
}
