% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_functions.R
\name{get_nPep_prMatrix}
\alias{get_nPep_prMatrix}
\title{Get the numer of razor/unique peptides per proteinGroup per sample}
\usage{
get_nPep_prMatrix(
  pr_matrix,
  id_column = "Protein.Group",
  peptide_level = "Stripped.Sequence"
)
}
\arguments{
\item{pr_matrix}{The report.pr_matrix file}

\item{id_column}{The IDs used in the output file}

\item{peptide_level}{The column specifying at which level intensities are
aggregated. See \link{summarize_peptide_intensities}}
}
\value{
a matrix with the count of razor/unique peptides
}
\description{
Get the numer of razor/unique peptides per proteinGroup per sample
}
\examples{
mat <- get_nPep_prMatrix(report.pr_matrix)

}
