% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_coverage.R
\name{get_peptide_position_in_protein}
\alias{get_peptide_position_in_protein}
\title{Find the amino acid locations of peptides within a protein sequence}
\usage{
get_peptide_position_in_protein(peptide_sequences, full_sequence)
}
\arguments{
\item{peptide_sequences}{A character vector with the peptide sequences.}

\item{full_sequence}{A character vector with the sequence of the parent protein.}
}
\value{
A data frame with the peptide sequences, their start position and end
position in the parent protein.
}
\description{
Find the amino acid locations of peptides within a protein sequence
}
\examples{

pep_sequence <- 'AARTGGGPLR'
full_sequence <- 'ISIFQKRSRWCAARTGGGPLRDPEIYLWWW'
positions <- get_peptide_position_in_protein(pep_sequence, full_sequence)

}
