% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{get_ranked_genes}
\alias{get_ranked_genes}
\title{Rank proteins/genes based on fold change}
\usage{
get_ranked_genes(res)
}
\arguments{
\item{res}{Results from get_DEPresults(). See details.}
}
\value{
A ranked, named vector with FC values and gene names.
}
\description{
Rank proteins/genes based on fold change
}
\details{
The res object should be obtained by running get_DEPresults() with a one-vs-one
(manual) comparison, otherwise it does not know which FC column to use to rank the proteins.
}
\examples{
\dontrun{
se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, condition1 = 'motif1', condition2 = 'neg_ctrl', type = 'manual')
GOBP <- get_genesets(db, collection = 'c5', subcollection = 'GOBP')
ranked_genes <- get_ranked_genes(res)

}
}
