% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibaq_peptides}
\alias{ibaq_peptides}
\title{List with calculated theoretical tryptic peptides for mouse and human for iBAQ
Peptides are calculated using 'cleaver' package with 'trypsin' set as enzyme,
zero miscleavages, and peptide length 7-30.}
\format{
\subsection{\code{ibaq_peptides}}{

A list of length 2
}
}
\source{
Uniprot
}
\usage{
ibaq_peptides
}
\description{
List with calculated theoretical tryptic peptides for mouse and human for iBAQ
Peptides are calculated using 'cleaver' package with 'trypsin' set as enzyme,
zero miscleavages, and peptide length 7-30.
}
\keyword{datasets}
