% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{load_msigdb}
\alias{load_msigdb}
\title{Downloads MSigDB}
\usage{
load_msigdb(organism = "hs")
}
\arguments{
\item{organism}{Specifies for which organism to download data. 'hs' for human, 'mm' for mouse.}
}
\value{
A GSEAbase object with genesets
}
\description{
Downloads MSigDB
}
\examples{
db <- load_msigdb('mm') #downloads data for mouse.
}
