% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_imputation.R
\name{mixed_imputation}
\alias{mixed_imputation}
\title{Perform mixed imputation on summarizedExperiment object}
\usage{
mixed_imputation(se, cutoff = "empirically")
}
\arguments{
\item{se}{The summarizedExperiment object}

\item{cutoff}{The cutoff used for MAR/MNAR classification. See \link{create_imputation_mask}}
}
\value{
A summarizedExperiment object with complete cases after mixed imputation.
}
\description{
Perform mixed imputation on summarizedExperiment object
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign, report.pr_matrix, impute = 'none')
se <- mixed_imputation(se)
}
