% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_imputation.R
\name{mixed_imputation_matrix}
\alias{mixed_imputation_matrix}
\title{Perform mixed imputation on a data matrix}
\usage{
mixed_imputation_matrix(data, conditions, cutoff = "empirically")
}
\arguments{
\item{data}{A matrix with intensity values}

\item{conditions}{Character vector specifying the experimental conditions}

\item{cutoff}{The cutoff used for MAR/MNAR classification. See \link{create_imputation_mask}}
}
\value{
A matrix without missing values
}
\description{
Perform mixed imputation on a data matrix
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign, report.pr_matrix, impute = 'none')
data_missing <- as.matrix(SummarizedExperiment::assay(se)) # Intensity matrix with missing values
conditions <- unique(se$conditions) # The experimental conditions
data_full <- mixed_imputation_matrix(data_missing, conditions, cutoff = 'empirically')
}
