% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{perform_GSEA}
\alias{perform_GSEA}
\title{Perform GSEA based on DEP results}
\usage{
perform_GSEA(res, genesets)
}
\arguments{
\item{res}{Results from get_DEPresults(). See details.}

\item{genesets}{MSigDB gene sets retrieved with get_genesets()}
}
\value{
A data frame with GSEA results
}
\description{
Perform GSEA based on DEP results
}
\details{
The res object should be obtained by running get_DEPresults() with a one-vs-one
(manual) comparison, otherwise it does not know which FC column to use to rank the proteins.
}
\examples{
\dontrun{
se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, condition_1 = 'motif1', condition2 = 'neg_ctrl', type = 'manual')
GOBP <- get_genesets(db, 'c5', 'GOBP')
gsea <- perform_GSEA(res, GOBP)
}

}
