% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_imputation.R
\name{perform_mixed_imputation}
\alias{perform_mixed_imputation}
\title{Perform mixed imputation over a matrix containing data of a single condition.}
\usage{
perform_mixed_imputation(matrix, matrix_masks)
}
\arguments{
\item{matrix}{A matrix with intensities. Should contain all replicates of
a single condition.}

\item{matrix_masks}{Boolean masks returning from \link{create_imputation_mask}}
}
\value{
A matrix with all complete observations after imputation.
}
\description{
Perform mixed imputation over a matrix containing data of a single condition.
}
\examples{
\dontrun{
masks <- create_imputation_mask(data_matrix)
data <- perform_mixed_imputation(data_matrix, masks)
}
}
