% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plots.R
\name{plot_gsea_barplot}
\alias{plot_gsea_barplot}
\title{Plots GSEA results as bar plot.}
\usage{
plot_gsea_barplot(
  gsea,
  pos_color = "gold1",
  neg_color = "darkblue",
  padj_cutoff = 0.05,
  top_n = Inf,
  remove_prefix = F,
  max_name_length = Inf,
  break_names = T
)
}
\arguments{
\item{gsea}{Output from \link{perform_GSEA}.}

\item{pos_color}{Character vector specifying which color to use for activated pathways.}

\item{neg_color}{Character vector specifying which color to use for repressed pathways.}

\item{padj_cutoff}{The maximum p.adjust value allowed for inclusion of the pathway.}

\item{top_n}{The maximum number of pathways included. Takes the top_n pathways
with the lowest p.adj values.}

\item{remove_prefix}{Boolean specifying to remove the prefix from pathway names.}

\item{max_name_length}{Numeric value specifying the max length of pathway names.}

\item{break_names}{Boolean value that indicates if pathway names should be shown
on two lines when the length exceedds the max_name_length parameter}
}
\value{
A ggplot bar plot object
}
\description{
Plots GSEA results as bar plot.
}
\examples{
\dontrun{
gsea <- perform_GSEA(res, genesets)

# Bar plot of 10 pathways with lowest padj values.
barplot <- plot_gsea_barplot(gsea, top_n = 10)

#Bar plot of 10 pathways with lowest values and different colors for bars.
barplot <- plot_gsea_barplot(gsea, pos_color = 'red3', neg_color = 'dodgerblue', top_n = 10)
}
}
