% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plots.R
\name{plot_gsea_dotplot}
\alias{plot_gsea_dotplot}
\title{Plots GSEA data as a faceted dot plot showing NES, padj, and set size.}
\usage{
plot_gsea_dotplot(
  gsea,
  padj_cutoff = 0.05,
  top_n = Inf,
  remove_prefix = F,
  max_name_length = Inf,
  break_names = T
)
}
\arguments{
\item{gsea}{gsea Output from \link{perform_GSEA}.}

\item{padj_cutoff}{padj_cutoff The maximum p.adjust value allowed for inclusion of the pathway.}

\item{top_n}{The maximum number of pathways included. Takes the top_n pathways
with the lowest p.adj values.}

\item{remove_prefix}{Boolean specifying to remove the prefix from pathway names.}

\item{max_name_length}{Numeric value specifying the max length of pathway names.}

\item{break_names}{Boolean value that indicates if pathway names should be shown
on two lines when the length exceedds the max_name_length parameter}
}
\value{
A ggplot object.
}
\description{
Plots GSEA data as a faceted dot plot showing NES, padj, and set size.
}
\examples{
\dontrun{
gsea <- perform_GSEA(res, genesets)
dotplot <- plot_gsea_dotplot(gsea, top_n = 10)
}
}
