% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plots.R
\name{plot_gsea_enrichment_plot}
\alias{plot_gsea_enrichment_plot}
\title{Creates individual GSEA enrichment plots for selected pathwways.}
\usage{
plot_gsea_enrichment_plot(
  genesets,
  pathways,
  res,
  line_color = "green",
  ES_color = "red",
  linewidth = 1,
  ticksSize = 0.2
)
}
\arguments{
\item{genesets}{A list with pathways/genesets created with \link{get_genesets}.}

\item{pathways}{A character vector specifying which pathways need to be plotted.}

\item{res}{Results from \link{get_DEPresults}. Requires a 1-vs-1 comparison.}

\item{line_color}{Character specifying which color the solid line in the plot is.}

\item{ES_color}{Character specifying which colro the dashed line specifying ES-score in the plot is.}

\item{linewidth}{Numeric value specifying the thickness of the plotted lines.}

\item{ticksSize}{Numeric value specifying the thickness of the ticks in lower part graph.}
}
\value{
An single or list of ggplot2 objects.
}
\description{
Creates individual GSEA enrichment plots for selected pathwways.
}
\examples{
\dontrun{
db <- load_msigdb()
genesets <- get_genesets(db, 'h')
res <- get_DEPresults(se, 'motif1', 'neg_ctrl')
pathways <- c('HALLMARK_MYC_TARGETS_V1', 'HALLMARK_TGF_BETA_SIGNALING')
enrichment_plots <- plot_gsea_enrichment_plot(genesets, pathways, res)
}

}
