% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_coverage.R
\name{plot_protein_coverage}
\alias{plot_protein_coverage}
\title{Plots the coverage of proteins.}
\usage{
plot_protein_coverage(
  se,
  genes,
  positions = NULL,
  zoom = NULL,
  fasta = NULL,
  organism = "hs",
  combine_overlap = T,
  dodge_labels = T,
  scaling = "centered"
)
}
\arguments{
\item{genes}{The gene names of the proteins that you would like to plot.}

\item{positions}{A numeric vector indiciting the amino acid positions that
you want to highlight.}

\item{zoom}{A numeric with two values (start and stop) indicating the specific
part of the protein that you want to plot.}

\item{fasta}{Data frame with protein information when not using a default option.
See \link{get_ibaq_peptides}}

\item{organism}{Specifies which default protein database to use.
'hs' for human, or 'mm' for mouse.}

\item{combine_overlap}{Boolean specifying whether to combine miscleaved peptides.}

\item{dodge_labels}{Boolean specifying wheter position labels should be plotted
on different heights so that they do not overlap in the plot.}

\item{scaling}{Boolean value indicating whether the peptide intensities should
be centered over the different replicates.}

\item{pr_matrix}{The SummarizedExperiment object from \link{create_se}}
}
\value{
A ggplot2 object showing the found peptides for a protein.
}
\description{
Plots the coverage of proteins.
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign, report.pr_matrix)
smad4 <- plot_protein_coverage(se, 'SMAD4', positions = c(100, 150))

# Zoom in on first 200 AA of protein
smad4 <- plot_protein_coverage(se, 'SMAD4', positions = c(100, 150), zoom = c(1, 200))
}
