% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_upset}
\alias{plot_upset}
\title{Plots upset plot showsing overlapping significant proteins between comparisons}
\usage{
plot_upset(res, ..., comparisons = "all", names = NULL)
}
\arguments{
\item{res}{A data frame with results from get_DEPresults()}

\item{...}{Additional results data frames (optional)}

\item{comparisons}{A character vector specifying which comparisons
to include. The default 'all' considers all comparisons present in res.}

\item{names}{(optional). Specifies the labels for the different subsets.
Defaults to the comparison names in colnames(res)}
}
\value{
A ggplot object
}
\description{
Plots upset plot showsing overlapping significant proteins between comparisons
}
\details{
The layout of the resulting upset plot can be easily edited to your liking.
Check the vignette of ggupset to see how individual components can be altered.
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, type = 'all')
vd <- plot_upset(res)

}
