% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_venn_diagram}
\alias{plot_venn_diagram}
\title{Plots a Venn diagram showing overlapping significant proteins}
\usage{
plot_venn_diagram(
  res,
  ...,
  comparisons = "all",
  colors = c("#b3e2cd", "#fdcdac", "#cbd5e8", "#f4cae4", "#e6f5c9"),
  names = NULL
)
}
\arguments{
\item{res}{A data frame with results from get_DEPresults()}

\item{...}{Additional results data frames (optional)}

\item{comparisons}{A character vector specifying which comparisons
to include. The default 'all' considers all comparisons present in res.}

\item{colors}{The colors used for the Venn diagram.}

\item{names}{(optional). Specifies the labels for the different subsets.
Defaults to the comparison names in colnames(res)}
}
\value{
A Venn diagram plot
}
\description{
Plots a Venn diagram showing overlapping significant proteins
}
\examples{

se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, type = 'all')
plot_venn_diagram(res) # Default option
plot_venn_diagram(res, colors = c('green', 'blue', 'red')) # Uses non-default
# colors
plot_venn_diagram(res, comparisons = c('neg_ctrl_vs_motif1',
                                       'neg_ctrl_vs_motif2')) # Only
                                       # includes two comparisons


}
