% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVolcano}
\alias{plotVolcano}
\title{create volcano plots for all comparisons present in the results file}
\usage{
plotVolcano(
  res,
  pval_cutoff = 0.05,
  fc_cutoff = 1,
  label = "sig",
  top_n = NULL,
  up_color = "red3",
  down_color = "dodgerblue",
  ns_color = "grey70"
)
}
\arguments{
\item{res}{A data frame with results from get_DEPresults()}

\item{pval_cutoff}{The p.adj significance cutoff}

\item{fc_cutoff}{The log2 fold change significance cutoff}

\item{label}{Specifies which points to label. The default is 'sig', labeling
all significant points. Entering a value for top_n limits the labeling to
the top_n up- and top_n down-regulated proteins based on the p.adj.
When providing a vector with protein names, only those points are labeled.}

\item{top_n}{Specifies how many significant points to label.}

\item{up_color}{Specifies the color of the significantly
up-regulated proteins.}

\item{down_color}{Specifies the color of the significantly down-regulated
proteins.}

\item{ns_color}{Specifies the color of non-significant proteins.}
}
\value{
A single ggplot object (1 comparison) or a list with ggplot objects.
}
\description{
create volcano plots for all comparisons present in the results file
}
\examples{

library(ggplot2)
library(ggrepel)

se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, type = 'all')
vol <- plotVolcano(res, top_n = 10) # Labels the top 10 upregulated and
# top10 downregulated proteins based on fdr.
vol <- plotVolcano(res, label = c('SMAD3', 'SMAD4')) # Only labels SMAD3/4
vol <- plotVolcano(res, up_color = 'green', down_color = 'yellow') #
# Gives a very ugly volcano plot
}
