% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepare_barplot_data}
\alias{prepare_barplot_data}
\title{Prepare data for plot_DEP_barplot function.}
\usage{
prepare_barplot_data(res, comparisons = "all", names = NULL)
}
\arguments{
\item{res}{A data frame with results from get_DEPresults()}

\item{comparisons}{A character vector specifying which comparisons
to include. The default 'all' considers all comparisons present in res.
The bars are plotted in the order of specified comparisons.}

\item{names}{(optional). Specifies the labels for the different bars.
Defaults to the comparison names in colnames(res)}
}
\value{
A data frame with the number of significant proteins per condition.
}
\description{
Prepare data for plot_DEP_barplot function.
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, type  = 'all')
data <- prepare_barplot_data(res) # All conditions
data <- prepare_barplot_data(res, comparisons = c('neg_ctrl_vs_motif1',
                                                  'neg_ctrl_vs_motif2')) #
# Only two comparisons included.

}
