% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_se.R
\name{prepare_diann_data}
\alias{prepare_diann_data}
\title{Tidies sample names and parses an experimental design.}
\usage{
prepare_diann_data(pg_matrix, pr_matrix)
}
\arguments{
\item{pg_matrix}{The report.pg_matrix file}

\item{pr_matrix}{The report.pr_matrix file}
}
\value{
A list with a tidied pg_matrix, pr_matrix, and parsed expDesign.
}
\description{
Tidies sample names and parses an experimental design.
}
\details{
This functions tidies sample names and prepares and experimental design,
assuming that the structure of the sample names is <prefix_condition_replicate>.
An example of the prepared sample names and expDesign is printed in the console.
}
\examples{
tidy_data <- prepare_diann_data(report.pg_matrix, report.pr_matrix)
pg_matrix <- tidy_data$pg_matrix
pr_matrix <- tidy_data$pr_matrix
}
