% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plots.R
\name{prepare_gsea_data}
\alias{prepare_gsea_data}
\title{Prepares data for the different GSEA visualization options}
\usage{
prepare_gsea_data(gsea, padj_cutoff = 0.05, top_n = Inf)
}
\arguments{
\item{gsea}{Output from \link{perform_GSEA}.}

\item{padj_cutoff}{The maximum p.adjust value allowed for inclusion of the pathway.}

\item{top_n}{The maximum number of pathways included. Takes the top_n pathways
with the lowest p.adj values.}
}
\value{
A data frame with filtered GSEA data.
}
\description{
Prepares data for the different GSEA visualization options
}
\examples{
\dontrun{

gsea <- perform_GSEA(res, genesets)
data <- prepare_gsea_data(gsea) # all significant pathways
data <- prepare_gsea_data(gsea, top_n = 10) # 10 pathways with lowest p.adjust values
(minimally padj < 0.05)
}
}
