% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepare_MA_data}
\alias{prepare_MA_data}
\title{Prepare the data for making MA plots}
\usage{
prepare_MA_data(
  res,
  label = NULL,
  ns_color = "grey70",
  up_color = "red3",
  down_color = "dodgerblue"
)
}
\arguments{
\item{res}{A data frame with results from get_DEPresults()}

\item{label}{Specifies which points to label. The default is 'sig', labeling
all significant points. Entering a value for top_n limits the labeling to
the top_n up- and top_n down-regulated proteins based on the p.adj.
When providing a vector with protein names, only those points are labeled.}

\item{ns_color}{Specifies the color of non-significant proteins.}

\item{up_color}{Specifies the color of the significantly
up-regulated proteins.}

\item{down_color}{Specifies the color of the significantly down-regulated
proteins.}
}
\value{
A list with data frames.
}
\description{
Prepare the data for making MA plots
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign)
se <- add_median_peptide_intensity(se, report.pr_matrix)
res <- get_DEPresults(se, type = 'all')
data <- prepare_MA_data(res)
}
