% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_coverage.R
\name{prepare_peptide_data}
\alias{prepare_peptide_data}
\title{Prepares the data for protein coverage plots.}
\usage{
prepare_peptide_data(
  pr_matrix,
  genes,
  positions = NULL,
  zoom = NULL,
  fasta = NULL,
  organism = "hs",
  combine_overlap = T,
  dodge_labels = T
)
}
\arguments{
\item{pr_matrix}{The report.pr_matrix file.}

\item{genes}{The gene names of the proteins that you would like to plot.}

\item{positions}{A numeric vector indiciting the amino acid positions that
you want to highlight.}

\item{zoom}{A numeric with two values (start and stop) indicating the specific
part of the protein that you want to plot.}

\item{fasta}{Data frame with protein information when not using a default option.
See \link{get_ibaq_peptides}}

\item{organism}{Specifies which default protein database to use.
'hs' for human, or 'mm' for mouse.}

\item{combine_overlap}{Boolean specifying whether to combine miscleaved peptides.}

\item{dodge_labels}{Boolean specifying wheter position labels should be plotted
on different heights so that they do not overlap in the plot.}
}
\value{
A list containing a data frame with information about the tryptic
peptides, and a data frame with the AA positions that are highlighted.
}
\description{
Prepares the data for protein coverage plots.
}
\examples{
coverage_data <- prepare_peptide_data(report.pr_matrix,  'SMAD4', positions = c(100, 120))

}
