% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP.R
\name{recode_sig_col}
\alias{recode_sig_col}
\title{Relabel the significance column in the DEP results data frame}
\usage{
recode_sig_col(res, pval_cutoff = 0.05, fc_cutoff = 1)
}
\arguments{
\item{res}{Results data frame after DEP analysis.}

\item{pval_cutoff}{The p.adj significance cutoff}

\item{fc_cutoff}{The absolute log2 fold change significance cutoff}
}
\value{
The results data frame with adjusted significance column
}
\description{
Relabel the significance column in the DEP results data frame
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, type = 'all')
res <- recode_sig_col(res, pval_cutoff = 0.01, fc_cutoff = 1)
}
