% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report.pg_matrix}
\alias{report.pg_matrix}
\title{Output file of DIANN with raw proteinGroup intensities on protein level.}
\format{
\subsection{\code{report.pg_matrix}}{

A data frame with 5948 rows and 13 columns.
}
}
\source{
241220_V118_LK_DNA_PD
}
\usage{
report.pg_matrix
}
\description{
Contains 9 samples with raw intensities after analysis with DIANN.
}
\keyword{datasets}
