% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report.pr_matrix}
\alias{report.pr_matrix}
\title{Output file of DIANN with raw intensities on peptide level.}
\format{
\subsection{\code{report.pr_matrix}}{

A data frame with 97329 rows and 19 columns.
}
}
\source{
241220_V118_LK_DNA_PD
}
\usage{
report.pr_matrix
}
\description{
Contains 9 samples with raw peptide intensities after analysis with DIANN.
}
\keyword{datasets}
