% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_functions.R
\name{summarize_peptide_intensities}
\alias{summarize_peptide_intensities}
\title{Combine different peptide variants into a single intensity}
\usage{
summarize_peptide_intensities(pr_matrix, peptide_level = "Stripped.Sequence")
}
\arguments{
\item{pr_matrix}{The report.pr_matrix file}

\item{peptide_level}{The column in pr_matrix that is used to aggregate.}
}
\value{
a matrix with a single peptide intensity per sample for each entry in the peptide_level column
}
\description{
Peptide intensities can be split out over different entries based on
different charge states or modifications. This function combines these
variants into one intensity per peptide sequence.
}
\examples{

mat <- summarize_peptide_intensities(report.pr_matrix)
}
