% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_se.R
\name{use_dep}
\alias{use_dep}
\title{Prepares summarizedExperiment object to work with functions from DEP package.}
\usage{
use_dep(se)
}
\arguments{
\item{se}{The summarizedExperiment object created with \link{prepare_se}}
}
\value{
A summarizedExperiment object.
}
\description{
Prepares summarizedExperiment object to work with functions from DEP package.
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign, report.pr_matrix)
se_dep <- use_dep(se)
}
