% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_functions.R
\name{add_median_peptide_intensity}
\alias{add_median_peptide_intensity}
\title{Adds median peptide intensities to summerizedExperiment object}
\usage{
add_median_peptide_intensity(se, peptide_intensities)
}
\arguments{
\item{se}{SummerizedExperiment object returned from DEP::make_se()}

\item{peptide_intensities}{Peptide intensities returned from
get_median_intensities_prMatrix.}
}
\value{
a summerizedExperiment object with peptide intensities as extra assay
}
\description{
Adds median peptide intensities to summerizedExperiment object
}
\examples{

se <- prepare_se(report.pg_matrix, expDesign)
mpi <- get_median_intensities_prMatrix(report.pr_matrix)
se <- add_median_peptide_intensity(se, mpi)
}
