% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{load_msigdb}
\alias{load_msigdb}
\title{Downloads MSigDB}
\usage{
load_msigdb(organism = "hs", id = "SYM")
}
\arguments{
\item{organism}{Specifies for which organism to download data. 'hs' for human, 'mm' for mouse.}

\item{id}{a character, representing the ID type to use ("SYM" for gene symbols and "EZID" for Entrez IDs).}
}
\value{
A GSEAbase object with genesets
}
\description{
Downloads MSigDB
}
\examples{
\dontrun{
db <- load_msigdb('mm') #downloads data for mouse.
db <- load_msigdb('hs', 'EZID') #downloads data for human with EnsembleID
}
}
