% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_se.R
\name{prepare_se}
\alias{prepare_se}
\title{Prepare summarizedExperiment object from diann report.pg_matrix file}
\usage{
prepare_se(
  pg_matrix,
  expDesign,
  pr_matrix = NULL,
  missing_thr = 0,
  min_peptides = 0,
  impute = "knn",
  mixed_cutoff = "empirically",
  remove_contaminants = TRUE
)
}
\arguments{
\item{pg_matrix}{the report.pg_matrix file from DIANN}

\item{expDesign}{A data frame with the experimental design. Should contain at least 'label', 'condition', and 'replicate' columns.}

\item{pr_matrix}{Optional argument. If the report.pr_matrix file from DIANN is provided, peptide information will be added to output.}

\item{missing_thr}{Integer specifying which proteinGroups are filtered out based on missing values.}

\item{min_peptides}{An integer specifing the cutoff for razor/unique peptides. The default is 0.}

\item{impute}{Specifies which imputatation method to use (default: knn). No imputation is done when entering 'none'.
See details for options.}

\item{mixed_cutoff}{Either 'empirally' or a value between 0-1. For details, see \link{mixed_imputation}}

\item{remove_contaminants}{A logical value specifying if potential
contaminants should be removed from the pg_matrix.}
}
\value{
A summarized Experiment object
}
\description{
Prepare summarizedExperiment object from diann report.pg_matrix file
}
\details{
For standard imputation options, see ?DEP::impute. For mixed imputation, see \link{mixed_imputation}
}
\examples{
se <- prepare_se(report.pg_matrix,
                expDesign, missing_thr = 1,
                impute = 'knn') # creates se with missing values imputed

se <- prepare_se(report.pg_matrix,
                expDesign,
                 missing_thr = 1,
                 impute = 'none') # creates se without imputing missing values
}
