% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepare_volcano_data}
\alias{prepare_volcano_data}
\title{Prepare data for volcano plots}
\usage{
prepare_volcano_data(
  res,
  pval_cutoff = 0.05,
  fc_cutoff = 1,
  label = "sig",
  top_n = NULL,
  up_color = "red3",
  down_color = "dodgerblue",
  ns_color = "grey70"
)
}
\arguments{
\item{res}{Data frame with results from get_DEPresults.}

\item{pval_cutoff}{The p.adj significance cutoff.}

\item{fc_cutoff}{The log2 fold change significance cutoff.}

\item{label}{Specifies which points to label. The default is 'sig', labeling
all significant points. Entering a value for top_n limits the labeling to
the top_n up- and down-regulated proteins based on the p.adj.
When providing a vector with protein names, onlythose points are labeled.}

\item{top_n}{Specifies how many significant points to label.}

\item{up_color}{Specifies the color of the significantly
up-regulated proteins.}

\item{down_color}{Specifies the color of the significantly down-regulated
proteins.}

\item{ns_color}{Specifies the color of non-significant proteins.}
}
\value{
A list with a data frame for each comparison present in res.
}
\description{
Prepare data for volcano plots
}
\examples{
se <- prepare_se(report.pg_matrix, expDesign)
res <- get_DEPresults(se, type = 'all')
data <- prepare_volcano_data(res)

}
